A36487 V 0.1.3
HV Disable From Customer will now cause the Pulse Sync Board to inhibit
pulses.  (Not when the customer disables XRay because that is handled in
hardware)

Changed the way that the can module signals that it has received a sync
command because that was subject to interrupt bashing.

Move as much code as possible out of interrupt so that we do not
inadvertently cause a logic error

Flashing LEDs at startup no longer inhibit program flow.  The can module
will be up and running in this state.  The LEDs will flash until the
startup time completes & the board has been initialized

Removed the testing/debugging code